+++ CHANGELOG.md
@@ -109,5 +109,8 @@
 * Accelerate speedMode using tXml in a separate worker for pre-computing time-spanning elements with @startid/@endid attributes
 * Support for toggling spiccato articulation with Shift+P
 # 0.3.3 patch
+* Add 'Show MEI Guidelines' in Help menu
+* Implement 'Consult Guidelines' feature ('alt-.')
+# 0.3.4 patch
 * Settings panel supporting the majority of Verovio options
 * Settings panel supporting CodeMirror options